refactor: FastAPI ミドルウェア設定を関数へ切り出し
